requirements.txt
streamlit==1.28.1
langchain==0.0.354
langchain-community==0.0.10
ollama==0.1.7
chromadb==0.4.18
assemblyai==0.17.0
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3
sentence-transformers==2.2.2
pydub==0.25.1